+++ src/App.tsx
@@ -1,8 +1,9 @@
+import { useContext } from 'react';
 import { Provider } from 'react-redux';
 import { Routes, Route, BrowserRouter } from 'react-router-dom';
 
 import { store } from './redux/store';
-import { ThemeProvider } from './context/ThemeContext';
+import { ThemeContext } from './context/ThemeContext';
 
 import MainPage from './pages/MainPage/MainPage';
 import NotFoundPage from './pages/NotFoundPage/NotFoundPage';
@@ -12,20 +13,22 @@ import Card from './components/Card/Card';
 import './App.css';
 
 const App: React.FC = () => {
+  const { theme } = useContext(ThemeContext);
+
   return (
     <Provider store={store}>
-      <ThemeProvider>
-        <BrowserRouter>
-          <ErrorBoundary>
+      <BrowserRouter>
+        <ErrorBoundary>
+          <div className={`wrapper ${theme}-theme`} data-testid="app-wrapper">
             <Routes>
               <Route path="/" element={<MainPage />}>
                 <Route path="person" element={<Card />} />
               </Route>
               <Route path="*" element={<NotFoundPage />} />
             </Routes>
-          </ErrorBoundary>
-        </BrowserRouter>
-      </ThemeProvider>
+          </div>
+        </ErrorBoundary>
+      </BrowserRouter>
     </Provider>
   );
 };
