+++ src/app/[[...slug]]/client.tsx
@@ -2,8 +2,13 @@
 
 import dynamic from 'next/dynamic';
 
+const ThemeProvider = dynamic(() => import('../../context/ThemeContext'));
 const App = dynamic(() => import('../../App'), { ssr: false });
 
 export function ClientOnly() {
-  return <App />;
+  return (
+    <ThemeProvider>
+      <App />
+    </ThemeProvider>
+  );
 }
