+++ src/context/ThemeContext.tsx
@@ -10,8 +10,8 @@ interface Props {
 
 const ThemeContext = createContext<Props>({
   theme: 'light',
-  setTheme: (theme: Theme) => {
-    console.log(`${theme} theme is set`);
+  setTheme: () => {
+    return;
   },
 });
 
@@ -33,4 +33,5 @@ const ThemeProvider: React.FC<{ children: React.ReactNode }> = ({
   );
 };
 
-export { ThemeProvider, ThemeContext };
+export { ThemeContext };
+export default ThemeProvider;
